fix(repo): migrate verified refactors to recoverable transaction
